pyproject.toml
[project]
name = "langchainproj2"
version = "0.1.0"
description = "MedlinePlus Healthcare RAG Assistant with Agentic and Non-Agentic modes"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.5",
    "langchain-openai>=1.1.0",
    "langchain-community>=0.4.0",
    "openai>=1.0.0",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",
    "faiss-cpu>=1.7.4",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
]
